pyproject.toml

[project]
name = "tech-trends-harvester"
version = "0.6.0"
description = "Find blog-worthy tech topics with Reddit posts, Google Trends, phrase extraction, and question detection."
readme = "README.md"
requires-python = ">=3.9"
authors = [{name = "Rich Lewis", email = "1149213+RichLewis007@users.noreply.github.com"}]
dependencies = [
    "PySide6>=6.9,<7",
    "requests>=2.32,<3",
    "feedparser>=6.0.12,<7",
    "pandas>=2.3,<3",
    "PyYAML>=6.0.3,<7",
    "python-dateutil>=2.9,<3",
    "beautifulsoup4>=4.14,<5",
    "praw>=7.8,<8",
    "pytrends>=4.9,<5",
    "lxml>=6.0,<7",
]

[tool.uv]
index-url = "https://pypi.org/simple"



